pyproject.toml
[project]
name = "turing-km"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "neo4j>=6.1.0",
    "networkx>=3.6.1",
    "openai>=2.31.0",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "pyvis>=0.3.2",
    "requests>=2.33.1",
]
